Initial implementation of queue using arrays
This file is a somewhat adaptation of the stack.c implementation. This is an extremely suboptimal implementation of queues using void*, which will result in a large number of memory spaces that became unusable if the queue increases a lot. Potential solutions involve the usage of the % mod operator (although these might be rather difficult), reset the queue whenever a realloc is needed (amortized O(1) solution), or just implementing the whole thing using linked lists (trivial).