Created queue implementation with linked list
Implementation of a queue using a linked list. Most of the operations were rather similar to those in the stack implementation, therefore it is very easy to do. Of note is to make sure that front doesn't end up deallocating the space that back is pointing to (when queue becomes 0 elements long). Further test to be done to make sure all the operations are working correctly.